Move ShowHelpEvent to events/controller
