+++ src/main/java/seedu/address/controller/Ui.java
@@ -15,7 +15,7 @@ import seedu.address.events.controller.JumpToListRequestEvent;
 import seedu.address.events.controller.PersonPanelSelectionChangedEvent;
 import seedu.address.events.storage.FileOpeningExceptionEvent;
 import seedu.address.events.storage.FileSavingExceptionEvent;
-import seedu.address.events.ui.ShowHelpEvent;
+import seedu.address.events.controller.ShowHelpEvent;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.person.ReadOnlyPerson;
@@ -23,7 +23,6 @@ import seedu.address.util.Config;
 import seedu.address.util.LoggerManager;
 
 import java.io.File;
-import java.util.List;
 import java.util.logging.Logger;
 
 /**
